refactor: rename internal direction name Error to Failure
When the actions were failed inside a workflow such as returning error,
the internal direction for this situation was Error.

But using Error as a direction is ambiguous with the real Error.
Also, in developing a workflow-representing DSL, "error" can be collided
with the golang-native "error" keyword.

rename "Error" direction keyword into "Failure"